requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Werkzeug==3.0.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
gunicorn
gevent==23.9.1
python-dotenv==1.0.1
